Move profile discovery call
